fixed ave

+++ src/main/cpp/t05_ave.cpp
@@ -22,7 +22,7 @@
 using namespace std;
 
 int t05_ave() {
-    float x, s = 0;
+    double x, s = 0;
     int c = 0;
     cin >> x;
     while(x!=0) {s = s + x; c++; cin >> x;}
